added built_in repo + fixed duplicates in case of successive function calls
