+++ Makefile
@@ -3,10 +3,10 @@
 #                                                         :::      ::::::::    #
 #    Makefile                                           :+:      :+:    :+:    #
 #                                                     +:+ +:+         +:+      #
-#    By: chonorat <chonorat@student.42lyon.fr>      +#+  +:+       +#+         #
+#    By: pgouasmi <pgouasmi@student.42lyon.fr>      +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/08/16 13:10:41 by chonorat          #+#    #+#              #
-#    Updated: 2023/08/17 16:41:31 by chonorat         ###   ########.fr        #
+#    Updated: 2023/08/18 13:19:53 by pgouasmi         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -20,7 +20,7 @@ _END = \033[0m
 _BOLD = \033[1m
 
 NAME = minishell
-CFLAGS = -Wall -Wextra -Werror
+CFLAGS = -Wall -Wextra -Werror -ggdb3
 RM = @rm -rf
 CC = @cc
 DIR = @mkdir -p
@@ -31,7 +31,10 @@ MAKE_LIBFT = @make -C Libft
 CLEAN_LIBFT = @make clean -C Libft
 FCLEAN_LIBFT = @make fclean -C Libft
 FILES = minishell\
-		Commands/env_cmd\
+		Commands/execution\
+		Commands/built_in/echo\
+		Commands/built_in/cd\
+		Commands/built_in/env\
 		Utils/free\
 		Utils/get_paths\
 		Utils/resources\
@@ -48,7 +51,7 @@ $(NAME): $(OBJS)
 
 Objects/%.o: Sources/%.c Makefile $(HEADER)
 	$(DIR) Objects
-	$(DIR) Objects/Commands Objects/Utils
+	$(DIR) Objects/Commands Objects/Utils Objects/Commands/built_in
 	$(PRINT) "Compiling ${_BOLD}$<$(_END)..."
 	$(CC) -c $(CFLAGS) $< -o $@
